+++ src/components/icons/Star.tsx
